Update zone options
With the addition of `text_color`, renaming the old, general `color` property to `fill_color` more clearly communicates their intended uses.

By referencing the `cityMaxZoom` variable inside ´minZoom´, we show how the two are related. The idea being that the zones become visible when the city fill layer is not.